Drawing again, using a *real* graph, now
Added an ID mechanism, which is super-simple, but works well enough

Made a WorldView, that works kinda-sorta, although I wonder about how it
would work in multiple panels and such. That will be the next thing I
test.

Nodes, Edges, and the World itself are all pretty much solid for now,
although I still need to add a containing mechanism to the Edges, or
build a separate layer on top of the World, to contain Entitys.